requirements.txt
Django==5.0.*
mysqlclient==2.2.4
python-decouple==3.8
Pillow==10.1.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7